Renaming testing router
