+++ webapp/src/tests/InGameChat.test.js
@@ -1,7 +1,7 @@
 import React, { act } from "react"
 import "@testing-library/jest-dom"
 import { render, screen, fireEvent, waitFor } from "@testing-library/react"
-import InGameChat from "../app/game/InGameChat"
+import InGameChat from "../components/game/InGameChat"
 import fetchMock from "jest-fetch-mock"
 
 // Mock fetch globally
